+++ src/client/renderer/api/camera/ICamera.hpp
@@ -18,53 +18,53 @@ public:
     virtual ~ICamera() = default;
 
     // 更新
-    virtual void update(f32 deltaTime) = 0;
+    virtual void update(f64 deltaTime) = 0;
 
     // 位置
-    virtual void setPosition(const glm::vec3& position) = 0;
-    virtual void setPosition(f32 x, f32 y, f32 z) = 0;
-    [[nodiscard]] virtual const glm::vec3& position() const = 0;
+    virtual void setPosition(const glm::dvec3& position) = 0;
+    virtual void setPosition(f64 x, f64 y, f64 z) = 0;
+    [[nodiscard]] virtual const glm::dvec3& position() const = 0;
 
     // 旋转（欧拉角，度）
-    virtual void setRotation(const glm::vec3& rotation) = 0;
-    virtual void setRotation(f32 pitch, f32 yaw, f32 roll = 0.0f) = 0;
-    [[nodiscard]] virtual const glm::vec3& rotation() const = 0;
+    virtual void setRotation(const glm::dvec3& rotation) = 0;
+    virtual void setRotation(f64 pitch, f64 yaw, f64 roll = 0.0f) = 0;
+    [[nodiscard]] virtual const glm::dvec3& rotation() const = 0;
 
     // 俯仰和偏航
-    [[nodiscard]] virtual f32 pitch() const = 0;
-    [[nodiscard]] virtual f32 yaw() const = 0;
-    [[nodiscard]] virtual f32 roll() const = 0;
+    [[nodiscard]] virtual f64 pitch() const = 0;
+    [[nodiscard]] virtual f64 yaw() const = 0;
+    [[nodiscard]] virtual f64 roll() const = 0;
 
-    virtual void setPitch(f32 pitch) = 0;
-    virtual void setYaw(f32 yaw) = 0;
-    virtual void setRoll(f32 roll) = 0;
+    virtual void setPitch(f64 pitch) = 0;
+    virtual void setYaw(f64 yaw) = 0;
+    virtual void setRoll(f64 roll) = 0;
 
     // 方向向量
-    [[nodiscard]] virtual glm::vec3 forward() const = 0;
-    [[nodiscard]] virtual glm::vec3 right() const = 0;
-    [[nodiscard]] virtual glm::vec3 up() const = 0;
+    [[nodiscard]] virtual glm::dvec3 forward() const = 0;
+    [[nodiscard]] virtual glm::dvec3 right() const = 0;
+    [[nodiscard]] virtual glm::dvec3 up() const = 0;
 
     // 移动
-    virtual void moveForward(f32 distance) = 0;
-    virtual void moveRight(f32 distance) = 0;
-    virtual void moveUp(f32 distance) = 0;
+    virtual void moveForward(f64 distance) = 0;
+    virtual void moveRight(f64 distance) = 0;
+    virtual void moveUp(f64 distance) = 0;
 
     // 旋转
-    virtual void rotate(f32 pitchDelta, f32 yawDelta) = 0;
-    virtual void look(f32 mouseDeltaX, f32 mouseDeltaY) = 0;
+    virtual void rotate(f64 pitchDelta, f64 yawDelta) = 0;
+    virtual void look(f64 mouseDeltaX, f64 mouseDeltaY) = 0;
 
     // 投影
     virtual void setProjectionMode(ProjectionMode mode) = 0;
-    virtual void setFOV(f32 fov) = 0;
-    virtual void setAspectRatio(f32 aspectRatio) = 0;
-    virtual void setNearFar(f32 nearPlane, f32 farPlane) = 0;
-    virtual void setOrthoSize(f32 size) = 0;
+    virtual void setFOV(f64 fov) = 0;
+    virtual void setAspectRatio(f64 aspectRatio) = 0;
+    virtual void setNearFar(f64 nearPlane, f64 farPlane) = 0;
+    virtual void setOrthoSize(f64 size) = 0;
 
     [[nodiscard]] virtual ProjectionMode projectionMode() const = 0;
-    [[nodiscard]] virtual f32 fov() const = 0;
-    [[nodiscard]] virtual f32 aspectRatio() const = 0;
-    [[nodiscard]] virtual f32 nearPlane() const = 0;
-    [[nodiscard]] virtual f32 farPlane() const = 0;
+    [[nodiscard]] virtual f64 fov() const = 0;
+    [[nodiscard]] virtual f64 aspectRatio() const = 0;
+    [[nodiscard]] virtual f64 nearPlane() const = 0;
+    [[nodiscard]] virtual f64 farPlane() const = 0;
 
     // 矩阵
     [[nodiscard]] virtual const glm::mat4& viewMatrix() const = 0;
@@ -76,12 +76,12 @@ public:
     [[nodiscard]] virtual const CameraConfig& config() const = 0;
 
     // 移动速度
-    virtual void setMoveSpeed(f32 speed) = 0;
-    [[nodiscard]] virtual f32 moveSpeed() const = 0;
+    virtual void setMoveSpeed(f64 speed) = 0;
+    [[nodiscard]] virtual f64 moveSpeed() const = 0;
 
     // 鼠标灵敏度
-    virtual void setMouseSensitivity(f32 sensitivity) = 0;
-    [[nodiscard]] virtual f32 mouseSensitivity() const = 0;
+    virtual void setMouseSensitivity(f64 sensitivity) = 0;
+    [[nodiscard]] virtual f64 mouseSensitivity() const = 0;
 
     // 脏标记
     [[nodiscard]] virtual bool isDirty() const = 0;
